Moved map entity updating/drawing to camera
