+++ src/Scene/MapTile.java
@@ -18,55 +18,12 @@ public class MapTile extends MapEntity {
         this.tileIndex = tileIndex;
     }
 
-    public MapTile(SpriteSheet spriteSheet, float x, float y, String startingAnimation, TileType tileType) {
-        super(x, y, spriteSheet, startingAnimation);
-        this.tileType = tileType;
-    }
-
-    public MapTile(float x, float y, HashMap<String, Frame[]> animations, String startingAnimation, TileType tileType) {
-        super(x, y, animations, startingAnimation);
-        this.tileType = tileType;
-    }
-
-    public MapTile(BufferedImage image, float x, float y, String startingAnimation, TileType tileType) {
-        super(image, x, y, startingAnimation);
-        this.tileType = tileType;
-    }
-
-    public MapTile(BufferedImage image, float x, float y, TileType tileType) {
-        super(image, x, y);
-        this.tileType = tileType;
-    }
-
-    public MapTile(BufferedImage image, float x, float y, TileType tileType, float scale) {
-        super(image, x, y, scale);
-        this.tileType = tileType;
-    }
-
-    public MapTile(BufferedImage image, float x, float y, TileType tileType, float scale, ImageEffect imageEffect) {
-        super(image, x, y, scale, imageEffect);
-        this.tileType = tileType;
-    }
-
-    public MapTile(BufferedImage image, float x, float y, TileType tileType, float scale, ImageEffect imageEffect, Rectangle bounds) {
-        super(image, x, y, scale, imageEffect, bounds);
-        this.tileType = tileType;
-    }
-
     public TileType getTileType() {
         return tileType;
     }
 
-    public void setTileType(TileType tileType) {
-        this.tileType = tileType;
-    }
-
     public int getTileIndex() {
         return tileIndex;
-    }
-
-    public void setTileIndex(int tileIndex) {
-        this.tileIndex = tileIndex;
     }
 
     public void update() {
